- Done valid & exception of vaccine & vaccinDetail

+++ Back-end/Child-Vaccine-Schedule-Tracking-System/src/main/java/org/gr1fpt/childvaccinescheduletrackingsystem/service/VaccineService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,9 +21,13 @@ public class VaccineService {
         return vaccineRepository.findAll();
     }
 
+    private String generateId() {
+        long count = vaccineRepository.count();
+        return "VX" + String.format("%05d", count + 1);
+    }
+
     public Vaccine create(Vaccine vaccine) {
-        if (vaccineRepository.existsById(vaccine.getVaccineId()))
-            throw new CustomException("Vaccine ID:" + vaccine.getVaccineId() + " Exist", HttpStatus.BAD_REQUEST);
+        vaccine.setVaccineId(generateId());
         return vaccineRepository.save(vaccine);
     }
 
@@ -31,11 +36,14 @@ public class VaccineService {
     }
 
     public void delete(String id) {
-        vaccineRepository.deleteById(id);
+        if(!vaccineRepository.existsById(id))
+            throw new CustomException("Vaccine ID:" + id + " Not Found", HttpStatus.BAD_REQUEST);
+        else
+            vaccineRepository.deleteById(id);
     }
 
     public Vaccine active(String id) {
-        Vaccine vaccine = vaccineRepository.findById(id).orElseThrow(() -> new CustomException("Child ID: " + id + " does not exist", HttpStatus.BAD_REQUEST));
+        Vaccine vaccine = vaccineRepository.findById(id).orElseThrow(() -> new CustomException("Vaccine ID: " + id + " does not exist", HttpStatus.BAD_REQUEST));
         if (vaccine.isActive())
             vaccine.setActive(false);
         else
@@ -62,5 +70,15 @@ public class VaccineService {
             throw new CustomException("Price not valid", HttpStatus.BAD_REQUEST);
         else
             return vaccineRepository.findByPriceBetween(min, max);
+    }
+
+    public List<Vaccine> findByAge (int ageMin, int ageMax){
+       List<Vaccine> list = new ArrayList<>();
+       for(Vaccine v : vaccineRepository.findAll()) {
+           if (v.getAgeMax() <= ageMax && v.getAgeMin() >= ageMin) {
+               list.add(v);
+           }
+       }
+       return list;
     }
 }
